Register declaratively and remove the root entry-point class
Move the remaining procedural registration out of the root
SemanticExternalQueryLookup class into the namespace. The extension
function is retained (it must run setup-time PHP for the class_alias
calls, which have no declarative equivalent) but relocated to
SEQL\Setup::onExtensionFunction, reading seqlgExternalRepositoryEndpoints
directly from the populated config global.

- Add SEQL\Setup with the class_alias calls (SMWExternalQueryLookup,
  SMWExternalAskQueryLookup) and the HookRegistry registration.
- Point extension.json ExtensionFunctions at SEQL\Setup::onExtensionFunction
  and drop AutoloadClasses; the class now autoloads via AutoloadNamespaces.
- Drop load_composer_autoloader: there is no composer autoload block and
  no third-party runtime dependency, so there is nothing local to load.
- Drop the SEQL_VERSION constant and SemanticExternalQueryLookup::getVersion();
  the test bootstrap now reads the version from extension.json.
- Delete the root SemanticExternalQueryLookup.php.
